+++ MapboxCoreNavigation/Assets.xcassets/reroute-sound.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "data" : [
+    {
+      "idiom" : "universal",
+      "filename" : "reroute-sound.mp3"
+    }
+  ]
+}
