+++ MapboxNavigation/String.swift
@@ -11,29 +11,3 @@ extension String {
         }
     }
 }
-
-
-extension String {
-    typealias Replacement = (of: String, with: String)
-
-    func byReplacing(_ replacements: [Replacement]) -> String {
-        return replacements.reduce(self) { $0.replacingOccurrences(of: $1.of, with: $1.with) }
-    }
-    
-    var addingXMLEscapes: String {
-        return byReplacing([
-            ("&", "&amp;"),
-            ("<", "&lt;"),
-            ("\"", "&quot;"),
-            ("'", "&apos;")
-            ])
-    }
-    
-    var removingPunctuation: String {
-        return byReplacing([
-            ("(", ""),
-            (")", ""),
-            ("_", "")
-            ])
-    }
-}
